Move tests to a tests dir
