+++ setup.cfg
@@ -35,6 +35,7 @@ universal = True
 [isort]
 profile = black
 force_single_line = true
+known_first_party = octal8
 
 [flake8]
 exclude = venv
@@ -52,3 +53,6 @@ strict_equality = true
 warn_redundant_casts = true
 warn_unused_ignores = true
 warn_unreachable = true
+
+[mypy-tests.*]
+disallow_untyped_defs = false
