+++ tox.ini
@@ -3,4 +3,4 @@ envlist = py37,py38,py39,py310
 
 [testenv]
 deps = -rrequirements.txt
-commands = pytest octal8.py
+commands = pytest -v tests/
